chore(app): extract out config parsing
This change extracts out the frontend configuration parsing functions
into a separate app-utils package, to prepare making the old RHDH
dynamic frontend configuration available to the new frontend system
configuration bridge when it is implemented.
