+++ packages/app-utils/.lintstagedrc.json
@@ -1,0 +1,4 @@
+{
+  "*": "prettier --ignore-unknown --write",
+  "*.{js,jsx,ts,tsx,mjs,cjs}": "backstage-cli package lint --fix"
+}
