+++ packages/app-utils/package.json
@@ -1,14 +1,26 @@
 {
-  "name": "app",
-  "version": "1.0.1",
-  "private": true,
-  "bundled": true,
+  "name": "@red-hat-developer-hub/app-utils",
+  "version": "1.0.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.esm.js",
+    "module": "dist/index.esm.js",
+    "types": "dist/index.d.ts"
+  },
   "backstage": {
-    "role": "frontend"
+    "role": "common-library",
+    "supported-versions": "1.0.0",
+    "pluginId": "app-utils",
+    "pluginPackage": "@red-hat-developer-hub/app-utils",
+    "pluginPackages": [
+      "@red-hat-developer-hub/app-utils"
+    ]
   },
   "scripts": {
-    "start": "janus-cli package start",
-    "build": "janus-cli package build",
+    "build": "backstage-cli package build",
     "clean": "backstage-cli package clean",
     "test": "backstage-cli package test --passWithNoTests --coverage",
     "lint:check": "backstage-cli package lint",
@@ -18,58 +30,30 @@
     "prettier:fix": "prettier --ignore-unknown --write ."
   },
   "dependencies": {
-    "@backstage-community/plugin-rbac-common": "1.18.0",
-    "@backstage/app-defaults": "1.6.2",
     "@backstage/catalog-model": "1.7.4",
     "@backstage/config": "1.3.2",
     "@backstage/core-app-api": "1.17.0",
-    "@backstage/core-components": "0.17.2",
     "@backstage/core-plugin-api": "1.10.7",
-    "@backstage/integration-react": "1.2.7",
     "@backstage/plugin-api-docs": "0.12.7",
     "@backstage/plugin-catalog": "1.30.0",
-    "@backstage/plugin-catalog-common": "1.1.4",
-    "@backstage/plugin-catalog-graph": "0.4.19",
     "@backstage/plugin-catalog-import": "0.13.0",
     "@backstage/plugin-org": "0.6.39",
-    "@backstage/plugin-permission-react": "0.4.34",
     "@backstage/plugin-scaffolder": "1.31.0",
-    "@backstage/plugin-scaffolder-react": "1.16.0",
-    "@backstage/plugin-search": "1.4.26",
-    "@backstage/plugin-search-react": "1.9.0",
-    "@backstage/plugin-user-settings": "0.8.22",
-    "@backstage/theme": "0.6.6",
-    "@emotion/react": "11.14.0",
-    "@emotion/styled": "11.14.1",
-    "@internal/plugin-dynamic-plugins-info": "*",
-    "@mui/icons-material": "5.18.0",
     "@mui/material": "5.18.0",
-    "@mui/styled-engine": "5.18.0",
-    "@red-hat-developer-hub/app-utils": "workspace:*",
-    "@red-hat-developer-hub/backstage-plugin-theme": "0.9.1",
-    "@red-hat-developer-hub/plugin-utils": "workspace:*",
-    "@scalprum/core": "0.8.3",
-    "@scalprum/react-core": "0.9.5",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-router-dom": "6.30.1",
-    "react-use": "17.6.0",
-    "tss-react": "4.9.19"
+    "lodash": "4.17.21"
   },
   "devDependencies": {
     "@backstage/cli": "0.32.1",
     "@backstage/test-utils": "1.7.8",
-    "@janus-idp/cli": "3.6.1",
-    "@scalprum/react-test-utils": "0.2.6",
     "@testing-library/dom": "9.3.4",
     "@testing-library/jest-dom": "6.6.4",
     "@testing-library/react": "14.3.1",
     "@testing-library/react-hooks": "8.0.1",
-    "@testing-library/user-event": "14.6.1",
     "@types/node": "22.16.3",
     "@types/react": "18.3.23",
     "@types/react-dom": "18.3.7",
     "prettier": "3.6.2",
+    "react": "18.3.1",
     "typescript": "5.9.2"
   },
   "browserslist": {
@@ -85,8 +69,9 @@
     ]
   },
   "files": [
-    "dist",
-    "config.d.ts"
+    "dist"
   ],
-  "configSchema": "config.d.ts"
+  "peerDependencies": {
+    "react": "16.13.1 || ^17.0.0 || ^18.2.0"
+  }
 }
