+++ packages/app-utils/src/components/MenuIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './MenuIcon';
