+++ packages/app-utils/src/dynamic/index.ts
@@ -1,0 +1,3 @@
+export * from './config';
+export * from './routes';
+export * from './utils';
