+++ packages/app-utils/src/dynamic/routes/index.ts
@@ -1,0 +1,1 @@
+export * from './bindAppRoutes';
