+++ packages/app-utils/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@backstage/cli/config/tsconfig.json",
+  "include": ["src"],
+  "exclude": ["node_modules"],
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "outDir": "../../dist-types/packages/app-utils",
+    "rootDir": "."
+  }
+}
