+++ packages/app/src/components/Root/ApplicationHeaders.tsx
@@ -2,10 +2,11 @@ import React, { useContext, useMemo } from 'react';
 
 import { ErrorBoundary } from '@backstage/core-components';
 
-import DynamicRootContext, {
+import {
   MountPoint,
   MountPointConfigBase,
-} from '@red-hat-developer-hub/plugin-utils';
+} from '@red-hat-developer-hub/app-utils';
+import DynamicRootContext from '@red-hat-developer-hub/plugin-utils';
 
 type Position = 'above-main-content' | 'above-sidebar';
 
@@ -29,7 +30,7 @@ export const ApplicationHeaders = ({ position }: { position: Position }) => {
 
   const appHeaderMountPoints = useMemo(() => {
     const appHeaderMP = (mountPoints['application/header'] ??
-      []) as ApplicationHeaderMountPoint[];
+      []) as unknown as ApplicationHeaderMountPoint[];
 
     return appHeaderMP.filter(({ config }) => config?.position === position);
   }, [mountPoints, position]);
