+++ packages/plugin-utils/src/types.ts
@@ -1,12 +1,5 @@
 import { Entity } from '@backstage/catalog-model';
 
-export type RouteBinding = {
-  bindTarget: string;
-  bindMap: {
-    [target: string]: string;
-  };
-};
-
 export type ResolvedDynamicRouteMenuItem =
   | {
       text: string;
@@ -44,29 +37,13 @@ export type ResolvedDynamicRoute = {
   };
 };
 
-export type MountPointConfigBase = {
+export type MountPointConfig = {
   layout?: Record<string, string>;
+  if: (e: Entity) => boolean;
   props?: Record<string, any>;
 };
 
-export type MountPointConfig = MountPointConfigBase & {
-  if: (e: Entity) => boolean;
-};
-
-export type MountPointConfigRawIf = {
-  [key in 'allOf' | 'oneOf' | 'anyOf']?: (
-    | {
-        [key: string]: string | string[];
-      }
-    | Function
-  )[];
-};
-
-export type MountPointConfigRaw = MountPointConfigBase & {
-  if?: MountPointConfigRawIf;
-};
-
-export type MountPoint = {
+export type ResolvedMountPoint = {
   Component: React.ComponentType<React.PropsWithChildren>;
   config?: MountPointConfig;
   staticJSXContent?:
@@ -79,16 +56,16 @@ export type EntityTabOverrides = Record<
   { title: string; mountPoint: string; priority?: number }
 >;
 
-export type MountPoints = Record<string, MountPoint[]>;
+export type MountPoints = Record<string, ResolvedMountPoint[]>;
 
-export type ScaffolderFieldExtension = {
+export type ResolvedScaffolderFieldExtension = {
   scope: string;
   module: string;
   importName: string;
   Component: React.ComponentType<{}>;
 };
 
-export type TechdocsAddon = {
+export type ResolvedTechdocsAddon = {
   scope: string;
   module: string;
   importName: string;
@@ -98,7 +75,7 @@ export type TechdocsAddon = {
   };
 };
 
-export type ProviderSetting = {
+export type ResolvedProviderSetting = {
   title: string;
   description: string;
   provider: string;
@@ -109,9 +86,9 @@ export type DynamicRootConfig = {
   entityTabOverrides: EntityTabOverrides;
   mountPoints: MountPoints;
   menuItems: ResolvedMenuItem[];
-  providerSettings: ProviderSetting[];
-  scaffolderFieldExtensions: ScaffolderFieldExtension[];
-  techdocsAddons: TechdocsAddon[];
+  providerSettings: ResolvedProviderSetting[];
+  scaffolderFieldExtensions: ResolvedScaffolderFieldExtension[];
+  techdocsAddons: ResolvedTechdocsAddon[];
 };
 
 export type ComponentRegistry = {
